Create Project 1- Explore Weather Trends
In this project, I analyzed local and global temperature data and compared the temperature trends where I live "Macca" to overall global temperature trends.

Extracting the data from the database in Udacity workspace I've exported the temperature data for the world as well as for the closest big cities to where I live, I chose Macca and to interact with the database I will use SQL.

Creating a line chart The chart line comparesMacca  temperatures with the global temperatures. then plotting the moving average rather than the yearly averages in order to smooth out the lines, making trends more observable.

Making observations All included in the pdf file, the observations include the similarities and differences between the world averages, Macca  averages, as well as overall trends.