Analyze A/B Test Results -Project 3
The project handles fictionary A/B test data and aims to deliver insight and recommendations if a fictionary new website should be implemented or not. The company has developed a new web page in order to try and increase the number of users who "convert," meaning the number of users who decide to pay for the company's product. 

My goal is to work through to help the company understand if they should implement the new page design, keep the old page, or perhaps run the experiment longer to make their decision.

Steps involved during the analysis were page assignment, removing duplicate user ids, hypothesis testing and standard statistical tests, and multiple regression modelling.